+++ examples/react/src/data-file.tsx
@@ -1,3 +1,4 @@
+import pino from 'pino';
 import React, { useRef } from 'react';
 import { Toolkit } from '@arcanejs/toolkit';
 
@@ -17,7 +18,14 @@ import path from 'path';
 
 const DATA_DIR = path.join(path.dirname(__dirname), 'data', 'data-file');
 
-const toolkit = new Toolkit();
+const toolkit = new Toolkit({
+  log: pino({
+    level: 'info',
+    transport: {
+      target: 'pino-pretty',
+    },
+  }),
+});
 
 toolkit.start({
   mode: 'automatic',