+++ examples/react/src/reorder-elements.tsx
@@ -1,9 +1,17 @@
+import pino from 'pino';
 import * as React from 'react';
 import { Toolkit } from '@arcanejs/toolkit';
 
 import { ToolkitRenderer, Button, Tab, Tabs } from '@arcanejs/react-toolkit';
 
-const toolkit = new Toolkit();
+const toolkit = new Toolkit({
+  log: pino({
+    level: 'info',
+    transport: {
+      target: 'pino-pretty',
+    },
+  }),
+});
 
 toolkit.start({
   mode: 'automatic',