+++ examples/react/src/tabs-colors-inputs.tsx
@@ -1,3 +1,4 @@
+import pino from 'pino';
 import * as React from 'react';
 import { Toolkit } from '@arcanejs/toolkit';
 
@@ -13,7 +14,14 @@ import {
   TextInput,
 } from '@arcanejs/react-toolkit';
 
-const toolkit = new Toolkit();
+const toolkit = new Toolkit({
+  log: pino({
+    level: 'info',
+    transport: {
+      target: 'pino-pretty',
+    },
+  }),
+});
 
 toolkit.start({
   mode: 'automatic',