+++ packages/react-toolkit/tsup.config.ts
@@ -1,9 +1,16 @@
 import { defineConfig } from 'tsup';
 
 export default defineConfig({
-  entry: ['src/index.ts', 'src/data.tsx'],
+  entry: ['src/index.ts', 'src/data.tsx', 'src/logging.ts'],
   format: ['cjs', 'esm'],
   splitting: true,
   dts: true,
-  external: ['@arcanejs/toolkit', 'lodash', 'react', 'react-reconciler', 'zod'],
+  external: [
+    '@arcanejs/protocol',
+    '@arcanejs/toolkit',
+    'lodash',
+    'react',
+    'react-reconciler',
+    'zod',
+  ],
 });