+++ packages/toolkit/.eslintrc.js
@@ -7,4 +7,14 @@ module.exports = {
     project: "./tsconfig.lint.json",
     tsconfigRootDir: __dirname,
   },
+  "overrides": [
+    {
+      "files": ["src/frontend/**/*"],
+      "rules": {
+        // This code is only used in the frontend, and the only code present
+        // Loggers aren't available here, so console logging is okay
+        "no-console": "off",
+      }
+    }
+  ]
 };