+++ packages/toolkit/src/backend/options.ts
@@ -2,6 +2,7 @@ import type * as http from 'http';
 
 import type { Application } from 'express';
 import { Server } from './server';
+import { Logger } from '@arcanejs/protocol/logging';
 
 export interface LightDeskOptions {
   /**
@@ -10,6 +11,15 @@ export interface LightDeskOptions {
    * This is important if a express server will be used that serves other paths.
    */
   path: string;
+  /**
+   * An optional object that can be used to output log events,
+   * we recommend using `pino` for this,
+   * as log levels etc... can be controlled.
+   *
+   * You can also always use `console` for logging,
+   * but this will be quite verbose.
+   */
+  log?: Logger;
 }
 
 export const DEFAULT_LIGHT_DESK_OPTIONS: LightDeskOptions = {
@@ -18,7 +28,15 @@ export const DEFAULT_LIGHT_DESK_OPTIONS: LightDeskOptions = {
 
 export type InitializationOptions =
   /** automatically start a simple  */
-  | { mode: 'automatic'; port: number }
+  | {
+      mode: 'automatic';
+      port: number;
+      /**
+       * Optional callback that is called when the server is ready,
+       * with the url that the server is running on.
+       */
+      onReady?: (url: string) => void;
+    }
   /** Create a websocket server that attaches to an existing express and http server */
   | { mode: 'express'; express: Application; server: http.Server }
   /** Create a websocket server that attaches to an existing express and http server */