+++ examples/react/CHANGELOG.md
@@ -1,5 +1,13 @@
 # @arcanejs/examples-react
 
+## 0.0.11
+
+### Patch Changes
+
+- Updated dependencies [9247670]
+  - @arcanejs/react-toolkit@0.4.0
+  - @arcanejs/toolkit@0.3.0
+
 ## 0.0.10
 
 ### Patch Changes