+++ packages/toolkit-frontend/CHANGELOG.md
@@ -1,5 +1,16 @@
 # @toolkit-frontend
 
+## 0.1.0
+
+### Minor Changes
+
+- 9247670: Remove console.log and introduce unified logging interface
+
+### Patch Changes
+
+- Updated dependencies [9247670]
+  - @arcanejs/protocol@0.2.0
+
 ## 0.0.3
 
 ### Patch Changes